Refactor auth test cases
