+++ tests/ballerina-integration-test/src/test/java/org/ballerinalang/test/auth/NoTokenPropagationTest.java
@@ -17,7 +17,7 @@
  */
 
 
-package org.ballerinalang.test.securelistener;
+package org.ballerinalang.test.auth;
 
 import org.ballerinalang.test.util.HttpClientRequest;
 import org.ballerinalang.test.util.HttpResponse;
@@ -28,18 +28,17 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Test cases for verifying token propagation scenario.
+ * Test cases for verifying no token propagation scenario.
  */
-@Test(groups = "secure-listener-test")
-public class TokenPropagationTest extends SecureListenerBaseTest {
+@Test(groups = "auth-test")
+public class NoTokenPropagationTest extends AuthBaseTest {
 
-    @Test(description = "With JWT Token propagation, authn success test")
+    @Test(description = "No JWT Token propagation, authn failure test")
     public void testTokenPropagationSuccess() throws Exception {
         Map<String, String> headers = new HashMap<>();
         headers.put("Authorization", "Basic aXN1cnU6eHh4");
-        HttpResponse response = HttpClientRequest.doGet(serverInstance.getServiceURLHttp(9094, "passthrough"),
-                headers);
+        HttpResponse response = HttpClientRequest.doGet(serverInstance.getServiceURLHttp(9190, "passthrough"), headers);
         Assert.assertNotNull(response);
-        Assert.assertEquals(response.getResponseCode(), 200, "Response code mismatched");
+        Assert.assertEquals(response.getResponseCode(), 401, "Response code mismatched");
     }
 }
