+++ tests/ballerina-integration-test/src/test/resources/auth/authservices/01_authn_config_inheritance_auth_disabling_test.bal
@@ -6,7 +6,7 @@ http:AuthProvider basicAuthProvider01 = {
     authStoreProvider:"config"
 };
 
-endpoint http:Listener listener {
+endpoint http:Listener listener01 {
     port:9090,
     authProviders:[basicAuthProvider01]
 };
@@ -19,7 +19,7 @@ endpoint http:Listener listener {
     }
 }
 
-service<http:Service> echo bind listener {
+service<http:Service> echo01 bind listener01 {
     @http:ResourceConfig {
         methods:["GET"],
         path:"/test",
