+++ tests/ballerina-integration-test/src/test/resources/auth/authservices/02_authn_config_inheritance_test.bal
@@ -7,7 +7,7 @@ http:AuthProvider basicAuthProvider02 = {
     authStoreProvider:"config"
 };
 
-endpoint http:Listener listener01 {
+endpoint http:Listener listener02 {
     port:9091,
     authProviders:[basicAuthProvider02]
 };
@@ -20,7 +20,7 @@ endpoint http:Listener listener01 {
     }
 }
 
-service<http:Service> echo01 bind listener01 {
+service<http:Service> echo02 bind listener02 {
     @http:ResourceConfig {
         methods:["GET"],
         path:"/test",
