+++ tests/ballerina-integration-test/src/test/resources/auth/authservices/05_service_level_auth_config_test.bal
@@ -7,7 +7,7 @@ http:AuthProvider basicAuthProvider05 = {
     authStoreProvider:"config"
 };
 
-endpoint http:Listener listener04 {
+endpoint http:Listener listener05 {
     port:9094,
     authProviders:[basicAuthProvider05]
 };
@@ -19,7 +19,7 @@ endpoint http:Listener listener04 {
         scopes:["scope2"]
     }
 }
-service<http:Service> echo04 bind listener04 {
+service<http:Service> echo05 bind listener05 {
     @http:ResourceConfig {
         methods:["GET"],
         path:"/test"
@@ -27,5 +27,14 @@ service<http:Service> echo04 bind listener04 {
     echo (endpoint caller, http:Request req) {
         http:Response res = new;
         _ = caller -> respond(res);
+    }
+
+    @http:ResourceConfig {
+        methods: ["GET"],
+        path: "/path/{id}"
+    }
+    path(endpoint caller, http:Request req, string id) {
+        http:Response res = new;
+        _ = caller->respond(res);
     }
 }
