--- tests/ballerina-integration-test/src/test/resources/secureListener/.ballerina/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
