--- tests/ballerina-integration-test/src/test/resources/secureListener/.gitignore
@@ -1,1 +1,0 @@
-target/
