--- tests/ballerina-integration-test/src/test/resources/secureListener/Ballerina.toml
@@ -1,4 +1,0 @@
-[project]
-org-name = "ballerina-test"
-version = "0.0.1"
-
