--- tests/ballerina-integration-test/src/test/resources/secureListener/ballerina.conf
@@ -1,9 +1,0 @@
-["b7a.users"]
-
-["b7a.users.isuru"]
-password="xxx"
-scopes="scope2"
-
-["b7a.users.ishara"]
-password="abc"
-scopes="scope1"
